% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argos.R
\name{argoSpeed}
\alias{argoSpeed}
\title{argoSpeed}
\usage{
argoSpeed(x)
}
\arguments{
\item{x}{a data.table containing tagID, locationDate,latitude, longitude.}
}
\value{
an updated data.table with speed as a new column.
}
\description{
argos speed along a track
}
\examples{
\dontrun{

argoSpeed(x)
} 
}
\author{
MV
}
