% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_kml.R
\name{hexBinary}
\alias{hexBinary}
\alias{kml.document}
\alias{kml.folder}
\alias{kml.placemark.line}
\alias{kml.placemark.points}
\alias{build.kml.folders}
\title{kml building blocks}
\usage{
hexBinary(col = "#A40000")

kml.document(title = "kml_file", body = "#---#")

kml.folder(name = "folder_1", body = "#---#")

kml.placemark.line(color = "#A40000", width = 3, name = "line1",
  coords = cbind(c(24.71, 24.71, 24.73), c(45.59, 45.61, 45.6)))

kml.placemark.points(color = "#4E9A06", scale = 1, name = "pt1",
  icon = "http://maps.google.com/mapfiles/kml/pal2/icon18.png",
  coords = c(24.71, 45.59), datetime = Sys.time())

build.kml.folders(dat, width = 3, scale = 1, colfun = col_factor("Paired",
  NULL))
}
\arguments{
\item{width}{line width}

\item{scale}{symbol scale}

\item{dat}{a data.table containing id, datetime_, lat, lon}

\item{colfun}{a function , default to scales::col_factor()}
}
\description{
builds kml folders using kml building blocks
}
\note{
for now dat is hardwired but this will change 
(dat should be a proper spacetime object or it should take id, datetime_, lat, lon as arguments)
}
\examples{
x = randomTracks() \%>\% build.kml.folders

}
