% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argos.R
\name{get_argos}
\alias{get_argos}
\title{get_argos}
\usage{
get_argos(con, tab = "argos", tagID = "ALL")
}
\arguments{
\item{con}{db connection}

\item{tab}{table name containing the data. should have tagID, locationDate,latitude, longitude}
}
\value{
data.table
}
\description{
loads argos data.
}
\examples{
    
\dontrun{

con = dbConnect(RSQLite::SQLite(), '~/argoSoap_scinam.sqlite' ) 
x =get_argos(con)
 }        

}
\author{
MV
}
