% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_kml.R
\name{kml}
\alias{kml}
\title{Export tracks to kml}
\usage{
kml(file = "~/Desktop/temp.kml", ...)
}
\arguments{
\item{...}{goes to build.kml.folders}

\item{A}{spatial points data.frame}
}
\value{
path to kml
}
\description{
Export tracks to kml. Each ID is kept in a separate kml _<Folder>_
}
\examples{
\dontrun{
kml(dat = randomTracks() )
 }

}
\author{
MV
}
