% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_kml.R
\name{kml_save}
\alias{kml_save}
\title{Export tracks to kml}
\usage{
kml_save()
}
\arguments{
\item{A}{spatial points data.frame}
}
\value{
path to kml
}
\description{
Export tracks to kml. Each ID is kept in a separate kml _<Folder>_
}
\examples{
    
\dontrun{
 }        

}
\author{
MV
}
